# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CoinVsa = typeof CoinVsa[keyof typeof CoinVsa];


export const CoinVsa = {
  ok_up: 'ok_up',
  ok_down: 'ok_down',
  manipulation_buy: 'manipulation_buy',
  manipulation_sell: 'manipulation_sell',
} as const;

export interface Coin {
  id: string;
  symbol: string;
  name: string;
  price: number;
  change24h: number;
  volume24h: number;
  followers: number;
  circulatingSupply: number;
  maxSupply: number;
  marketCap: number;
  vsa: CoinVsa;
}

export type VsaCheckStatus = typeof VsaCheckStatus[keyof typeof VsaCheckStatus];


export const VsaCheckStatus = {
  ok: 'ok',
  manipulation: 'manipulation',
} as const;

export interface VsaCheck {
  id: string;
  coinId: string;
  label: string;
  detail: string;
  status: VsaCheckStatus;
  volumeChange: number;
  priceChange: number;
}

export interface MarketSnapshot {
  updatedAt: string;
  coins: Coin[];
  vsaChecks: VsaCheck[];
  marketBreadth: number;
  sessionPnl: number;
}

export interface MarketCandle {
  /** Unix timestamp in seconds */
  time: number;
  open: number;
  high: number;
  low: number;
  close: number;
}

export type MarketCandlesInterval = typeof MarketCandlesInterval[keyof typeof MarketCandlesInterval];


export const MarketCandlesInterval = {
  '15m': '15m',
  '1h': '1h',
  '4h': '4h',
  '1d': '1d',
} as const;

export interface MarketCandles {
  coinId: string;
  interval: MarketCandlesInterval;
  candles: MarketCandle[];
}

export interface Strategy {
  id: string;
  name: string;
  shortName: string;
  category: string;
  timeframe: string;
  bias: string;
  rules: string[];
  winRate: number;
  trades: number;
}

export interface CopyBot {
  id: string;
  name: string;
  description: string;
  risk: string;
  status: string;
  strategyId: string;
}

export interface Watchlist {
  coinIds: string[];
}

export interface WatchlistInput {
  favorite: boolean;
}

export interface ExecutionLog {
  id: string;
  timestamp: string;
  type: string;
  message: string;
  coinId: string;
}

export type GetMarketCandlesParams = {
coinId: string;
interval: GetMarketCandlesInterval;
/**
 * @minimum 30
 * @maximum 100
 */
limit?: number;
};

export type GetMarketCandlesInterval = typeof GetMarketCandlesInterval[keyof typeof GetMarketCandlesInterval];


export const GetMarketCandlesInterval = {
  '15m': '15m',
  '1h': '1h',
  '4h': '4h',
  '1d': '1d',
} as const;