# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /market/candles:
    get:
      operationId: getMarketCandles
      tags: [market]
      summary: Get live Binance candlesticks for an asset
      parameters:
        - name: coinId
          in: query
          required: true
          schema: { type: string }
        - name: interval
          in: query
          required: true
          schema: { type: string, enum: [15m, 1h, 4h, 1d] }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 30, maximum: 100, default: 100 }
      responses:
        "200":
          description: Live exchange candlestick series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketCandles"
        "400":
          description: Invalid timeframe or request
        "404":
          description: Asset is not in the current live market universe
        "503":
          description: Exchange candlestick data is unavailable
...
  /watchlist/{coinId}:
    put:
      operationId: toggleWatchlist
      tags: [watchlist]
      summary: Set a coin favorite state
      parameters:
        - name: coinId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistInput"
      responses:
        "200":
          description: Updated favorite state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Watchlist"
...
components:
  schemas:
...
    MarketCandle:
      type: object
      required: [time, open, high, low, close]
      properties:
        time: { type: integer, description: Unix timestamp in seconds }
        open: { type: number }
        high: { type: number }
        low: { type: number }
        close: { type: number }
    MarketCandles:
      type: object
      required: [coinId, interval, candles]
      properties:
        coinId: { type: string }
        interval: { type: string, enum: [15m, 1h, 4h, 1d] }
        candles:
          type: array
          items: { $ref: "#/components/schemas/MarketCandle" }
...
    Watchlist:
      type: object
      required: [coinIds]
      properties:
        coinIds:
          type: array
          items: { type: string }
    WatchlistInput:
      type: object
      required: [favorite]
      properties:
        favorite: { type: boolean }
...